Statistics refactoring
