+++ ydb/core/statistics/aggregator/tx_aggr_stat_response.cpp
@@ -1,6 +1,6 @@
 #include "aggregator_impl.h"
 
-#include <ydb/core/statistics/stat_service.h>
+#include <ydb/core/statistics/service/service.h>
 
 namespace NKikimr::NStat {
