+++ ydb/core/statistics/aggregator/tx_response_tablet_distribution.cpp
@@ -1,7 +1,7 @@
 #include "aggregator_impl.h"
 
 #include <ydb/core/protos/hive.pb.h>
-#include <ydb/core/statistics/stat_service.h>
+#include <ydb/core/statistics/service/service.h>
 
 namespace NKikimr::NStat {
