+++ ydb/core/statistics/database/database.cpp
@@ -1,6 +1,6 @@
-#include "save_load_stats.h"
+#include "database.h"
 
-#include "events.h"
+#include <ydb/core/statistics/events.h>
 
 #include <ydb/library/table_creator/table_creator.h>
 #include <ydb/library/query_actor/query_actor.h>
