+++ ydb/core/statistics/database/ya.make
@@ -1,0 +1,19 @@
+LIBRARY()
+
+SRCS(
+    database.h
+    database.cpp    
+)
+
+PEERDIR(
+    ydb/core/base
+    ydb/core/engine/minikql
+    ydb/core/protos
+    ydb/core/tablet
+    ydb/core/tablet_flat
+    ydb/library/minsketch
+)
+
+YQL_LAST_ABI_VERSION()
+
+END()
