+++ ydb/core/statistics/ut/ut_statistics.cpp
@@ -3,7 +3,7 @@
 #include <ydb/library/actors/testlib/test_runtime.h>
 
 #include <ydb/core/statistics/events.h>
-#include <ydb/core/statistics/stat_service.h>
+#include <ydb/core/statistics/service/service.h>
 
 #include <ydb/public/sdk/cpp/client/ydb_result/result.h>
 #include <ydb/public/sdk/cpp/client/ydb_table/table.h>
