+++ ydb/core/statistics/ut/ya.make
@@ -24,8 +24,8 @@ SRCS(
     ut_common.cpp
     ut_aggregator.cpp
     ut_statistics.cpp
-    ut_save_load_stats.cpp
-#    ut_statistics_service.cpp
+    ut_database.cpp
+    #ut_service.cpp
 )
 
 END()
