+++ ydb/core/statistics/ya.make
@@ -1,11 +1,8 @@
 LIBRARY()
 
 SRCS(
+    common.h
     events.h
-    stat_service.h
-    stat_service.cpp
-    save_load_stats.h
-    save_load_stats.cpp
 )
 
 PEERDIR(
@@ -21,6 +18,8 @@ END()
 
 RECURSE(
     aggregator
+    database
+    service
 )
 
 RECURSE_FOR_TESTS(
