+++ ydb/core/testlib/basics/ya.make
@@ -21,6 +21,7 @@ PEERDIR(
     ydb/core/mind
     ydb/core/node_whiteboard
     ydb/core/quoter
+    ydb/core/statistics/service
     ydb/core/tablet_flat
     ydb/core/testlib/actors
     ydb/core/tx/columnshard
